[editor/shared Preview] Support rendering previews when items at positions < 0.
