+++ app/src/editor/shared/components/Preview/index.jsx
@@ -7,6 +7,8 @@ import ModulePreview from './ModulePreview'
 import styles from './preview.pcss'
 
 type Props = {
+    x: number,
+    y: number,
     aspect: {
         height: number,
         width: number,
@@ -26,37 +28,41 @@ const Preview = ({
     scale,
     viewBoxScale,
     ...props
-}: Props) => (
-    <div
-        className={cx(styles.root, className)}
-    >
-        <svg
-            preserveAspectRatio="none"
-            viewBox={`0 0 ${Math.round(aspect.width * viewBoxScale)} ${Math.round(aspect.height * viewBoxScale)}`}
-            {...props}
+}: Props) => {
+    const { x, y } = preview
+    const viewBox = `${x} ${y} ${Math.round(aspect.width * viewBoxScale)} ${Math.round(aspect.height * viewBoxScale)}`
+    return (
+        <div
+            className={cx(styles.root, className)}
         >
-            {children}
-            {preview.modules.map(({
-                height,
-                key,
-                x,
-                title,
-                y,
-                type,
-                width,
-            }) => (
-                <ModulePreview
-                    height={height * scale}
-                    key={key}
-                    title={title || ''}
-                    type={type}
-                    width={width * scale}
-                    x={x * scale}
-                    y={y * scale}
-                />
-            ))}
-        </svg>
-    </div>
-)
+            <svg
+                preserveAspectRatio="none"
+                {...props}
+                viewBox={viewBox}
+            >
+                {children}
+                {preview.modules.map(({
+                    height,
+                    key,
+                    x,
+                    title,
+                    y,
+                    type,
+                    width,
+                }) => (
+                    <ModulePreview
+                        height={height * scale}
+                        key={key}
+                        title={title || ''}
+                        type={type}
+                        width={width * scale}
+                        x={x * scale}
+                        y={y * scale}
+                    />
+                ))}
+            </svg>
+        </div>
+    )
+}
 
 export default Preview
